pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "google-genai>=1.73.1",
    "numpy<2",
    "onnxruntime==1.19.2",
    "protobuf<4",
    "sentence-transformers==2.7.0",
    "torch==2.2.2",
]
